Cleaned up README.  Add share function.
Started work on the share function.  This will be the function that
allows someone to share the REPL across the local network.  At first I
thought I would find some clever way to either take over the existing
REPL, or maybe proxy *in* and *out* but that proved to be difficult.
Instead, I think I'm just going to create a sub-REPL and provide it
the necessary hooks for broadcast.